requirements.txt
Flask==2.3.2
Flask-SocketIO==5.2.0
pycryptodome==3.17
ecdsa==0.17.0
streamlit==1.24.0
